Check for existence of test execution times for the minimum and maximum frequency tests in the CPU userspace governor tests.
If there is no execution time, the test never started, so don't try to
display results. Display an error message along with what was already done
to set the test as failed.

Additionally, if one of the times is unavailable, don't do statistical
analysis, since it may result in, at best, worthless data.