Changed file open mode so that FIFO opening always succeeds.
This is to avoid the situation where the backend fails to start and the 
frontend blocks forever trying to open (checkbox.lib.fifo.FifoBase),
because opening a FIFO on read-only or write-only mode blocks until there's 
something writing or reading on the other end. Setting O_NONBLOCK causes 
an open for reading to fail with SIGPIPE. This seems to work as per FIFO(7)
at least on Linux.

The problem of what to do if the backend doesn't respond (in which case 
the frontend also blocks waiting to read) still remains.