Change how we parse the devices argument
Using the 'reminder' feature of argparse is confusing as it will also
consume arguments like --help and will not respect choices keyword to
validate the parsed value. While the same effect could be used with
nargs="*", after discussion with the certification team we've decided to
remove this feature (allowing users to not pass any device type). Thus
instead of 'reminder' we use the nargs="+" keyword. This properly
validates each argument device type argument and disallows users to pass
an empty list.