checkbox.dbus.udisks2: Drop all DBus types early
DBus API has an annoying collection of types that make reading and
comparing data more difficult than it otherwise would. I found a few
cases where some of the code failed to compare a regular string to an
DBus string so I've decided to drop the DBus type signatures as early as
possible.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>