scripts/removable_storage_watcher: Summarize the device delta in verbose mode
This patch can be dropped if necessary. All it does is adds a summary
function that prints (via logging.info()) summary of all the changes in
each _on_change() call. The main purpose of this method is lightweight
debugging that is easier to understand as compared to the full debugging
log.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>