scripts/removable_storage_watcher: Replace the implementation of removable_storage_watcher
This is a large change. Instead of listening to all the detailed signals
and trying to understand when 'insert' or 'remove' had occurred. This
was not really optimal as the logic was complex and could not capture
everything we wanted (such as filesystem becoming mounted).

The old logic is entirely removed in favour of the new mehanism that
builds a delta between the initial (reference) objects exposed by
UDisks2 and the actual objects after each change.

The delta is then fed to a validator that has all of the "business
logic" and is easy to verify or modify. The code is also symmetrical,
behaving exactly the same for both insertions and removals.  The only
thing that differentiates this is the direction of the delta.

This code also has one TODO, related to USB 3.0 detection which is not
supported in UDisks2 and must be added later. I will first prototype a
solution for that as right now all I have are hints on where it is.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>