scripts: correct error conditions in md5_hash_file()
md5_hash_file() may misbehave if .read() raises an exception.
In such case the hash of a part of the file would be reported
without a hint of trouble.

In addition, if initial open() fails the result is confusingly, 1.

In both cases, on any error the function returns None and logs
the details of the problem that had occurred. Error handling is
also shorter than before.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>