scripts: fix ValueError from lookup_udev_device()
Commit a5c69ddf54efe57b64e5310248c8ea3e60cb1c4f labeled
"scripts: use lookup_udev_device in removable_storage_watcher"
has introduced the use of lookup_udev_device() function. That
function is intended to look up the udev device associated with
udisks object that is a block device. In the faulty commit however
we pass the drive udisks2 object instead of the object representing
the actual device we are inspecting at the time.

This patch looks up the UDisks2 object associated with object_path
(either past or present, depending on the type of interaction, removal,
or insertion), and passes that to lookup_udev_device.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>