scripts: lessen the mounted requirement only for empty ConnectionBus
Originally all storage devices are mounted upon insertion.
When udisks2 was introduced the auto-mounting policy was re-designed
and only devices that are typically "removable" were being mounted.
That includes usb, firewire and SD cards. This has left-out eSATA
as it has no discoverable differences as compared to plain SATA.

To accommodate that change in removable_storage_watcher whenever
the "" (empty) desired ConnectionBus property was detected (such
as when eSATA monitoring was requested on command line) we have removed
the mounted filesystem requirements.

This change was applied incorrectly, when multiple desired values
of ConnectionBus property are used, for example for both eSATA and USB
then removal detection would fail, as the USB drive was mounted when
it got removed.

To work around this problem, the mounted filesystem property is only
removed for drives with empty ConnectionBus. Checks in the same loop,
for other devices, retain the full set of tests.

Rudimentary testing on USB card reader shows the change works correctly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>