udevadm: Allow using UdevadmParser with a string
The stream interface to UdevadmParser is rather strange as the code
simply discards the stream in favour of a normal string early on.
This design is problematic with Unicode strings though as it requires
the user to setup an appropriate stream initially (using proper encoding)
or suffer confusing Unicode issues from the parser module.

In order to follow the rule of thumb to keep all bytes/unicode
translations at program boundaries this patch allows UdevadmParser to
be instantiated with a string argument. This allows the caller to handle
all string encoding and never expose the parser to byte buffers.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>