Add plainbox.impl.runner module
This module contains the implementation of the IJobRunner interface.

The JobRunner class has two public methods, the constructor and the
run_job(job) method. Everything else is an implementation detail.

There are multiple methods that implement running a job with
a particular plugin value. The _plugin_shell() and _plugin_resource()
being notable only implemented one.

Running actual commands is implemented in _run_command() with the
aid of _get_checkbox_env(). That part needs some more work to manage
implementing the io_log properly so that IO can be monitored in the UI
in real time.

The run_job() method also allows each _plugin_xxx() to return a TestResult
instance. This is something that may need generalizing so that things
like attachments can be managed sensibly.

There is no way to plug the UI in at all, this needs much more work to
be practical. There are no tests at all (apart from what gets executed
by other tests)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>