Add the plainbox.impl.box module
This module contains an attempt to implement a basic text user interface
and needs much more work to work properly (with the UI callbacks
and the rest of the APIs in a way that would scale to GTK or Qt-based UI)

It is also used as a simple way to implement the plainbox.public interface
but it may need to go away (as the public interface has not been reviewed
at all).

There are some basic smoke tests but nothing that actually runs jobs
yet. This part is a R&D component that allows me to understand how
the job runner should be implemented. As it matures it will influence
the public API design and the UI abstract base classes.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>