Add plainbox.impl.depmgr and assocated tests
This module implements the DependencySolver class with one class method
DependencySolver.resolve_dependencies() that takes a list of jobs
as argument and either throws one of the thee new exceptions or
returns a list of jobs that are safe to run in that order.

The three new exceptions Dependency{Cycle,Missing,Duplicate}Error
indicate that a dependency loop, unresolved dependency or duplicate job
name has been found, respectively. They also carry enough state for
informative error messages or error recovery.

The solver can also consider a subset of jobs to visit. This allows one
to select a few jobs that actually need to run while still taking
advantage of the validation and resource dependency resolution.

The module has 100% test coverage

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>