Rewrite the box module.
The introduction of the session module allowed me to remove virtually
all of the session logic from the UI. The new UI does more and is much
easier to understand. There are several new features:

1) Initial support for interactive jobs. There is a new command line
argument to run in non-interactive mode. Interactivity is also
disabled when stdin or stdout are not connected to a tty.

2) Full support for proper state handling: jobs with failed
dependencies won't be started. Proper human-readable error messages
are provided in all such cases.

3) Somewhat hacky but correct support for local jobs. While local jobs
are handled at the session level they don't cause the desired job
list to be re-computed (as only the UI knows the rules that select
or unselect jobs). This is something that should get more work to get
right. I'm sure the session module can do a better job at hiding that
complexity

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>