Change ResourceProgram.evaluate() to evaluate_or_raise()
This is a UI-driven change. When a requirement program fails
it is good to show the operator which part of the program
prevents a job from running. With the old interface this
required the UI layer to re-compute the same expressions just
to find out which part caused problems.

The new interface always returns True or raises an informative exception.
The exception has references to the failing expression and can be
converted to a human readable, descriptive error message.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>