Rewrite the runner module.
This is a large change and after a few attempts I realized that
splitting it up is pointless. There are many technical bits that got
changed but they all follow one basic new idea: to make the job runner
stateless.

A stateless job runner is far easier to test, it also allows me to move
all of the state handling (that was spread between runner.py and box.py)
into a new module - session.

Stateless runner is also no longer responsible for understanding job
output or checking if a particular jobs can start at a given time.
All of those tasks are now handled externally.

The usage of subprocess was replaced with the vendorized extcmd module.
This allows the runner to easily build data for the JobResutl.io_log
property. Using extcmd also allows us to expose process IO callbacks
to the UI layer. A default implementation of that is providedd when
no callbacks are passed to JobRunner.__init__(), which simply logs
everything to the debug logger.

The simplified runner can now run more job types (more "plugins")
notably apart from existing shell and resource jobs it now handles
local and manual jobs.

The introduction of manual jobs required some additional callbacks
to interactively get the outcome from the user. I'm pretty sure that this
interface is _wrong_ for any UI toolkit and will force us to have
a separate thread but this can be revisited when an actual toolkit
becomes considered. It will most likely involve setting up a way to share
events and IO descriptors with the UI toolkit event loop. This primitive
API allows me to build the console version and get moving quickly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>