Add support for saving a subset of session state after al tests
(this commit needs splitting)

There are three main parts:

1) New code in box.py where a SessionStateExporter classes are being
discovered and initialized. There are new associated command line
options --output-file --output-format and --output-options.

2) A piece of common infrasturcture in plainbox.impl.exporter
Each exporter has a method that computes a subset of state to
export (this part is mostly shared among various actual exporters) and
a dump method that saves that subset to a stream

3) A small collection of three exporters: as plain text, as json as as yaml
of execution

There are no tests yet, this code is just for early review

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>