make_job() create Origin that tracks source code location
The make_job() is a small utility function used in several tests.
Sometimes, when a job fails to parse or causes other problems, it
may be difficult to find where it came from.

This patch creates an Origin() object that points at the caller
of make_job() and associates it with the newly created test JobDefinition
instance

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>