box: use re.match istead of fnmatch
This will allow existing whitelist patterns to match succesfully since they assume that proper regular expressions are used. Also updated the tests to use regular expressions. To align with what checkbox did, and also to try and make use of -i intuitive, the patterns automatically have ^ prepended and $ appended, so that the pattern doesn't match part of a job name.