exporter: add xml exporter
[work in progress: needs attachments]

This exporter for SessionState data can create data that is compatible
with the Hardware Certification website. It is not a general-purpose
exporter, it depends on particular jobs to have executed to be useful.

Also, unlike other exporters, it has a constructor that accepts some
arguments (they all have defaults but we may need to define some
particular values in practice). Unlike all the exporters is also accepts
no options. For that reasons it is not installed in the 'selectable'
exporters and instead should be used explicitly somehow by the
application. So far it should be used by the `plainbox sru` commands.

This patch adds a dependency on lxml (python3-lxml) to use the RelaxNG
schema validator.

There are some 'unit tests' and a larger test that tries to export data
from a real session into XML. The data file was originally generated by
Sylvain on an earlier version of PlainBox. It should be carefully
updated if the internal format changes.

There are a few new files in test-data/xml-exporter, those all are for
making unit tests readable.