session: properly handle SessionState() called with duplicates
This is a part of fix to lp: #1157264

When SessionState() is initialized with a list containing duplicates
bad things happen. The error recovery code in update_desired_jobs()
is not prepared for any duplicates and things break in odd places.

To fix that, have SessionState reject duplicate jobs in the same way
that DependencySolver does. As a small exception to this rule, jobs
that are prefect clones are silently coalesced.

This makes the constructor of SessionState throw DependencyDuplicateError
exception when real problems have occurred and prevent any testing from
being started. Unlike local jobs that can also silently produce duplicate
jobs, having initial name clashes cannot be solved at this layer

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Reported-by: Sean Feole <sean.feole@canonical.com>