exporter: xml: generate output that can be read by checkbox parser.
This means not outputting test outcomes that are considered invalid by
checkbox. This was determined pretty much by trial-and-error, but
basically it meant mapping plainbox's 'not-supported' outcome, which
checkbox is unaware of, to 'skip'. A _STATUS_MAP is provided
for possible, future changes in this behavior, to ensure xml exporter
continues to produce checkbox-compatible output.