box: load config file on startup
This patch makes the plainbox command load and validate the
configuration file on startup.

Since particular command may or may not need valid configuration doing
this at this stage would be premature. Errors are silently ignored, that
is, they are not displayed to the user in any way but incorrect
configuration is _not_ applied.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>