runner: Change the runner behaviour when we run commands as another user
In order to apply the policy file to the trusted launcher, I had to remove
the call to env(1) to let the plainbox-cmd be argv[0] for pkexec.

The list of environment settings are passed a separate argument to the
trusted launcher.

The previous method is still in use to run the root commands when
checkbox "src" mode is detected