checkbox-old/scripts/gst_pipeline_scripts: Force the Gst sink device
The DisplayPort/HDMI port is not automatically selected when a monitor is
plugged using this interface.
By setting the PULSE_SINK environment variable, the Gstreamer sink
(i.e autoaudiosink) overrides the PulseAudio default sink which may be
different from the fallback sink set with pacmd.
The sink name is the first returned by "pacmd list-sinks" with contains the
HDMI pattern.