checkbox-old/checkbox/scripts/audio_settings.py: Fix the profile backup
The previous method made the assumption that the saved card profile should
be applied to card #0 when restoring the settings.
It's wrong, especially for system using multiple ALSA cards.

The fix saves the current profile for all cards and use their symbolic
names to restore the profiles.

Profiles configuration is backed up using configparser.