plainbox:dbus: add DBus utility module
This patch adds a subclass of standard dbus.service.Object that supports
properties via the new dbus.service.property decorator. This is based on
implementation of Get(), Set() and GetAll() methods and a new version of
the Introspect() method.

Currently it is not tested and does not support the PropertiesChanged
signal. The semantics of GetAll() might be incorrect in case of errors
when calling Get(). The way properties are discovered is redundant and
could be instead handled with _dbus_class_table.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>