plainbox:result: add IOLogRecordWriter and IOLogRecordReader
The two recorder reader and writer classes can be used to efficiently
write and read records to a text stream. They can be used to either
stream records to a file on disk or stream them from disk to memory.

In either case the whole sequence of IOLogRecord entries is never kept
in memory, read or written to disk as a whole.

This class is meant to be used with the new, upcoming, DiskJobResult
that is suitable for efficient processing of job results with arbitrarily
large log files

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>