plainbox:integration_tests: parametrize integration tests by scenario path, not job name
To support testing a single job multiple times (in various configurations)
we need to decouple the test scenario from the job it operates on.

Each job can have multiple scenarios that affect it, for example
with and without dependencies, on various OS versions, with various
hardware, etc.

NOTE: this patch breaks the tree but is shorter to review

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com