plainbox: split CheckBox class into dedicated 'deb' and 'src' providers
This patch mainly splits the CheckBox class into CheckBoxSrcProvider and
CheckBoxDebProvider, both being based on Provider1 class. There are a few
other places that still care about one or the other provider "modes"
that have not been converted to methods on the provider or job object.
Those instead check for object class (a but ugly but not much different
from checking a magic constant).

Both new providers are exported as entry points. There is a third 'auto'
provider that is not a class but a function that prefers the 'src'
provider but then falls back to the 'deb' provider, if required.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>