plainbox:impl:service, highlevel: Added callback when for job results
To start a new job, it's far better to wait for the previous one to send
its results over dbus. Relying on a job_state_map property changed was
no a good thing as many updates happen internally within PlainBox.

The RunningJob instance is now responsible to send a dbus signal when a
job result is ready to be added to the state map.