checkbox:parsers:tests:test_udevadm: Changes to test expectations from previous commit, since most systems sprouted a few AUDIO devices (that were previously OTHER or None)
Note that we're mostly just changing the quantities of expected devices,
since the categorization changes mean that, compared to the parser
overhaul from rev 2235, we are finding a few extra devices. mostly these
comprise audio and tty devices. A few of the expected devices changed
their index, so we update that as well.

None of the devices changed buses. Also, none of the "interesting"
device categories (except for AUDIO) changed their number of devices,
any device movement happened between OTHER and "None", per the linked
bug which requests assigning some devices to category: None.