plainbox:testing_utils: fix make_job() creating unexpectedly different jobs
make_job() was defaulting to a few None data entries such as
'requires' and 'depends' even if those were not specified by the caller.
This lead to make_job(name='foo') != JobDefinition({'name': 'foo'}) and
obviously to those two jobs having different checksums.

(This patch does not address the fact that make_job() has a default value
for plugin ("dummy") that is also making the inequality above true. Fixing
that requires larger patches to various tests that still assume 'dummy'
is there that I don't want to do right now.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>