plainbox:session: move legacy API to dedicated subclass of SessionState
This patch moves several methods and properties from SessionState into a
new subclass SessionStateLegacyAPIOrignalImpl (aliased as
SessionStateLegacyAPI) and adjusts all current usage of SessionState to
use that class instead. The legacy API is described by the
ISessionStateLegacyAPI interface.

The intent for this patch is to set the stage for the next patch which
adds another subclass of SessionState (which also implements the legacy
API interface) but this time using the new suspend/resume API and the
session manager class.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>