plainbox:session:storage: add legacy_mode argument to SessionStorage.create()
The new argument controls if sessions are created in "legacy"
single-session mode. In that mode at most one session can be created and
each new session replaces the predecessor. This is implemented using a
symlink that tracks the 'last-session'. In non-legacy mode arbitrary
number of sessions can be created and 'last-session' symlink is not
used.