plainbox:session:storage: add LockedStorageError exception
This exceptions is raised by save_checkpoint() when the temporary
(but with fixed name) 'next' file is encountered. Normally this file
should be removed during save_checkpoint() but an appropriate
signal can interrupt the process right after crating the file and before
renaming it to something more useful. While we should capture that
situation it is still possible to just kill -9 / crash the machine
and see that file left behind.

When this case is detected a new exception is raised, LockedStorageError,
to allow applications to call new APIs to forcibly unlock the session
and try again.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>