plainbox:session:legacy: add adapter using SessionManager via legacy API
This patch adds a new subclass of SessionState that implements all of
the legacy methods necessary to create, suspend, resume and destroy
sessions. The API is implemented using a mostly straightforward
pass-through to the SessionManager APIs except for the "create a new
session" vs "open an existing session" that uses delayed decision to
properly map that onto the SessionManager calls.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>