plainbox:service: replace - with _ in whitelist names.
This patch is a (temporary) workaround for whitelists that have - in their
names. The proper solution would be to stop using the name entirely but
that is somewhat blocked by the GUI component not doing whitelist
detection properly (by interfaces) but instead looks for certain patterns
in the object path.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>