plainbox:session:resume: add early_cb argument to SessionResumeHelper
This patch adds a callback to SessionResumeHelper.resume() that allows
that allows the caller to "see" the session object early, before the
bulk of resume operation happens.

The callback accepts one argument, a session which is being resumed.
This callback can be used to register signal listeners on the new
session before the resume() call returns.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>