plainbox:session:legacy: perform resume on legacy session
This patch changes how resume (from suspend/resume) is performed in
SessionStateLegacyAPICompatImpl. Instead of resuming a temporary session
and then copying the state across to the legacy session (which worked
but didn't fire correct signals) we now resume the actual legacy session
instance directly.

This is implemented by using the early_cb() callback to
SessionManager.load_session() to substitute the temporary (empty and
useless) session with the one we already have (the legacy instance being
resumed).

This has the advantage of being remarkably simple and correct. All of
the event listeners connected to the session are fired and no magic has
to be performed. The only downside is that the session is not "purged"
so you may see some of the old state being leaked (for example, when
resuming a session with less jobs than the one currently running) but in
practice that is not a problem and actually matches the (arguably
broken) semantics of resume() not being a class method in the legacy
API.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>