plainbox:session:state: adjust tests wrt JobResult.__eq__()
The tests we relying on the fact that SessionState.update_job_result()
was blindly assigning a new IJobResult over the previous one. With the
addition of proper JobResultBase.__eq__() this is no longer the case.

I was wondering if the test itself should be adjusted but I weighted on
just keeping it working the way it did before.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>