Added config files for offline tarball generation.
A directory contains the files needed by lp:dpkg_offline to create
driver-testing tarballs with all needed packages for offline
installation.

In addition to the repository adding script added by default by
dpkg_offline, this will add the start_testing convenience script and a
desktop file to make it easier to discover. The txt files contain
configuration on extra PPAs to add before downloading packages, and on
extra non-PPA files to add prior to packaging the tarball.

Note that this is now handled as a local apt repository so by relying on
apt-get for dependency resolution the process should be much smoother
than before.

A README file details the tarball creation process, which is done with
one command and should take less than 1 minute (though it depends on
link speed since package downloading is involved).