plainbox:session:resume: add SessionResumeHelper2
This patch adds a subclass of SessionResumeHelper1, SesionResumeHelper2,
that implements the support for additional entry in SessionMetaData,
app_blob. The patch is rather straightforward, subclassing the
'1' class and overriding _restore_SessionState_metadata to load
the new field. The only remark I would add is that I've decided not to
super-call the original method so that the logging call at the end
can really note when the whole meta-data was restored.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>