TestSelectionView.qml: Display test details in a popover element
Remove the test details banner to display test details using a popover.

The popup window will never be higher than half the size of the mainview.
The popup width is controlled by the contentWidth property set to the
mainview size minus 30 gu.

The popover content is scrollable, the flickableItem is defined to contain
a whole TestSelectionDetails item with only a vertical scrollbar.