TestSelectionView.qml: Replace Selection button by a Toolbar
This Toolbar is the default toolbar component of the Ubuntu SDK:
http://goo.gl/QYUCrH

To keep it always visible, the following properties are enabled:
- locked: true
- opened: true

Previous button are kept as is but the layout is a bit different now.

The Start Testing button is now orange (default color) and right aligned
(the default with ToolbarItems)

A new "info" button will open a ActionSelectionPopover list to display:
- test details
- test selection stats

The "info", "select/deselect all" buttons are left aligned using the "back"
property of the ToolbarItems component.

The info button color is lightAubergine to make the difference with the
select/deselect all buttons which are both coolgrey.