plainbox:session:storage: cope with 'last-session' being a broken symlink
The _remove_last_session() method was trying to properly reference
and remove the last session, both the actual session directory (which is
random each time and has to be followed through that symlink) as well
as the symlink itself.

This was failing when the symlink was broken as shutil.rmtree() raised
exceptions (rightfully so) that we didn't handle. Since we cannot really
do anything sensible there (in this case) it's okay to ignore that
exception

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>