LogViewer.qml: Use the WideDialog component instead of Dialog
The previous TextArea has been replaced by a TextEdit allowing a better
mouse selection. To restore the scrolling feature, the TextEdit is
included in a Flickable with Scrollbar attached.

The black Rectangle is used as a background (Terminal-like).

Note:
color: Qt.rgba(1, 1, 1, 0.9) is white, but a less aggressive version.
The same used internally by the SDK Dialog.