TestSelectionDetails.qml: Use the WideDialog component instead of Popover
All the content is included in a Flickable element.

The dialog height adjusts itself to the details height but the maximum
is defined to mainView.height / 1.618 (Golden ratio).
The dialog width is fixed to parent.width / 2 * 1.618.

The command TestSelectionDetailsItems text is now using the RichText format
to render text like a Terminal.