checkbox-old:parser:udevadm: Heuristic for Mediatek BT controllers.
This heuristic accounts for bluetooth devices which usually have
INTERFACE=224/*/1, however in the "field" we've run across a few
(Mediatek combo cards) that have unknown (255/255/255) values and thus
break the previous heuristic. We assume that if a device has these weird
INTERFACE values *but* it uses the btusb driver, then it must be a
bluetooth controller.  Other devices with btusb *but* with
INTERFACE=255/1/1 have been seen on systems where the actual usb
controller was identified by the old heuristic, so here we need to match
all three fields to avoid duplicating devices.

This applies only for devices we attempt to match through the
TYPE/INTERFACE rule, which normally is used only for USB devices. It's
placed at the very end of this block to only be used as a last resource.