checkbox-old:parsers:test_udevadm: Refactored the verify_devices method.
This change is needed to allow a device with no product name (presumably
because it's too new and the name is not yet in the database). The
format to specify expected devices for a system also changed slightly
(from a dict to a list of tuples). Existing device expectation blocks
were updated to match this new format.