plainbox-client: add the new plainbox client package
This package contains medium-to-long term code that will support a
pure-python implementation of the PlainBox DBus client API. It will
allow to create applications using plainbox that don't need to use the
low-level and inheriently unstable API.

In the future it can also hold implementations for languages other than
python, e.g. A Qt+C++ module.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>