gui-engine: Remove the title metadata check
First, it's not the best test to do to valid that saved sessions belong to
the GUI.

metadata are only available when the session is resumed and again why
resuming a session automatically if the tester didn't select his choice?

Before having a solid way to valid saved session, let's remove this code
to resume the session only when the tester decides it.