Added required resource jobs to autotesting whitelist.
This is the whitelist used in the mini CI loop. We removed all the
resource jobs because plainbox runs them "implicitly" when they are used
in a job's "requires" field. However, a few resources which are needed
for a valid submission are not requested by any job and so they weren't
being run.

I included a couple of resources that *are* used by jobs, but since they
are mandatory for a valid submission I thought it good to include them,
to signal the fact that these resources shouldn't be removed.

Also a "dummy" job that depends on those resources was added
(miscellanea/submission-resources)