plainbox:dbus:decorators: log exceptions inside dbus methods
This patch complements an earlier patch that confusingly should have
included this bit but only worked for properties. All methods
decorated with the improved plainbox @dbus.service.method decorator will
now log exceptions that happen inside, except for DBusExceptions which
are assumed to be raised explicitly with the intent to pass it to the
bus.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>